renamed files index.js, removed unnecessary imports
