+++ src/components/Footer/index.jsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import style from './Footer.module.css';
 import { Link } from 'react-router-dom';
 import logoImg from '../../assets/icons/logo.png';
