+++ src/components/LanguageSelect/index.jsx
@@ -1,6 +1,7 @@
 import { useState } from 'react';
 import { MenuItem, Select, FormControl, Box } from '@mui/material';
 import { styled } from '@mui/system';
+
 import UkraineFlag from '../../assets/flags/UkraineFlag.svg';
 import GermanyFlag from '../../assets/flags/GermanyFlag.svg';
 import USFlag from '../../assets/flags/USFlag.svg';
