+++ src/components/Layout/DefaultLayout.jsx
@@ -1,11 +1,9 @@
-import React from 'react';
 import Header from '../Header';
 import RegistrationPopup from '../RegistrationForm';
 import LoginForm from '../LoginForm';
 import Footer from '../Footer';
 
 export default function DefaultLayout({ children }) {
-  
     return (
         <>
             <Header />
