--- src/components/PersonMainSection/index.js
@@ -1,1 +1,0 @@
-export * from './PersonMainSection.jsx';
