+++ src/components/RegistrationForm/SuccessModalOpen/index.jsx
@@ -1,6 +1,5 @@
 import { Box, IconButton, Modal } from '@mui/material';
 import CloseIcon from '@mui/icons-material/Close';
-import React from 'react';
 import style from './SuccessModalOpen.module.css';
 
 export default function SuccessModalOpen({ openSuccessModal, setOpenSuccessModal }) {
